requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
idna==3.3
Pillow==9.1.0
PyJWT==2.3.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
